portal-impl/test/integration/com/liferay/portal/verify/VerifyResourcePermissionsTest.java
/**
 * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
 *
 * This library is free software; you can redistribute it and/or modify it under
 * the terms of the GNU Lesser General Public License as published by the Free
 * Software Foundation; either version 2.1 of the License, or (at your option)
 * any later version.
 *
 * This library is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
 * details.
 */

package com.liferay.portal.verify;

import com.liferay.portal.kernel.test.ExecutionTestListeners;
import com.liferay.portal.model.Company;
import com.liferay.portal.service.CompanyLocalServiceUtil;
import com.liferay.portal.test.LiferayIntegrationJUnitTestRunner;
import com.liferay.portal.test.MainServletExecutionTestListener;
import com.liferay.portal.util.CompanyTestUtil;

import org.junit.Test;
import org.junit.runner.RunWith;

/**
 * @author Manuel de la Peña
 */
@ExecutionTestListeners(listeners = {MainServletExecutionTestListener.class})
@RunWith(LiferayIntegrationJUnitTestRunner.class)
public class VerifyResourcePermissionsTest extends BaseVerifyTestCase {

	@Test
	public void testVerifyMoreThanOneCompany() throws Exception {
		verify(false);
	}

	@Test
	public void testVerifyOneCompany() throws Exception {
		verify(true);
	}

	@Override
	protected VerifyProcess getVerifyProcess() {
		return new VerifyResourcePermissions();
	}

	protected void verify(boolean oneCompany) throws Exception {
		Company company = null;

		if (!oneCompany) {
			company = CompanyTestUtil.addCompany();
		}

		doVerify();

		if (!oneCompany) {
			if (company != null) {
				CompanyLocalServiceUtil.deleteCompany(company);
			}
		}
	}

}